chore: move component to a folder
