+++ packages/frontend/src/components/common/Filters/FieldAutocomplete/FieldAutoComplete.styles.tsx
