+++ packages/frontend/src/components/common/Filters/index.tsx
@@ -22,8 +22,8 @@ import React, { FC, useCallback, useMemo } from 'react';
 import { useToggle } from 'react-use';
 import { v4 as uuidv4 } from 'uuid';
 import MantineIcon from '../MantineIcon';
-import FieldAutoComplete from './FieldAutoComplete';
-import { FieldAutoCompleteWrapper } from './FieldAutoComplete.styles';
+import FieldAutoComplete from './FieldAutocomplete/FieldAutoComplete';
+import { FieldAutoCompleteWrapper } from './FieldAutocomplete/FieldAutoComplete.styles';
 import FilterGroupForm from './FilterGroupForm';
 import { FieldWithSuggestions, useFiltersContext } from './FiltersProvider';
 import SimplifiedFilterGroupForm from './SimplifiedFilterGroupForm';
